+++ services/s3/src/test/java/software/amazon/awssdk/services/s3/checksums/S3ChecksumValidatingPublisherTest.java
@@ -33,6 +33,7 @@ import org.reactivestreams.Publisher;
 import org.reactivestreams.Subscriber;
 import org.reactivestreams.Subscription;
 import software.amazon.awssdk.core.checksums.Md5Checksum;
+import software.amazon.awssdk.services.s3.internal.checksums.S3ChecksumValidatingPublisher;
 import software.amazon.awssdk.utils.BinaryUtils;
 
 /**
